[css-flexbox] Don't always reuse cached layout for row flex containers
CalculateSizeBasedLayoutCacheStatusWithGeometry checks if the cached
layout information can be reused. The problem was that, if a flex
container had a percentage height, and the size of its containing block
changed, a new layout would only be performed in case it was a column
flex container. For row flex containers, the cached layout would be
reused.

However, row containers may also need a new layout in some cases.
This patch removes the check for column flex containers, so that row
ones can be included too.

Bug: 1103592

TEST=external/wpt/css/css-flexbox/height-percentage-with-dynamic-container-size.html

Change-Id: Ib2c119febdfa56b31f5524be3df85c9749dd1f6a